pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta" 

# For setuptools: https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
[project]
name = "aqua-core"
dynamic = ["version"]
authors = [ # AQUA core team reported here
    {name = "Paolo Davini", email = "paolo.davini@cnr.it"},
    {name = "Matteo Nurisso", email = "m.nurisso@isac.cnr.it"},
    {name = "Silvia Caprioli", email = "silvia.caprioli@polito.it"},
    {name = "Jost von Hardenberg", email = "jost.hardenberg@polito.it"},
    {name = "Marco Cadau", email = "marco.cadau@polito.it"},
    {name = "Emanuele Tovazzi", email = "emanuele.tovazzi@cnr.it"},
    {name = "Supriyo Ghosh", email = "supriyo.ghosh@bsc.es"},
    {name = "Maqsood Mubarak Rajput", email = "maqsoodmubarak.rajput@awi.de"}
]
description = "AQUA; a model evaluation framework for high-resolution climate model simulations"
readme = "README.md"
keywords = ["climate", "modeling", "diagnostics", "high-resolution"]
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython"
]
dependencies = [
    # binaries
    "eccodes==2.41.0",
    "netcdf4",
    "h5py>=3.12.1",
    # shared with conda
    "boto3",
    "cartopy>=0.24.1",
    "cfgrib", #enable support for GRIB files
    "dask[distributed]",
    "fastparquet", #enable support for parquet files
    "filelock",
    "gcsfs", # Google Cloud Storage File System: required by ARCO
    "intake==0.7.0", # Version 2 is a major refactor still in development
    "intake-xarray",
    "jinja2",
    "kerchunk",
    "matplotlib",
    "metpy",
    "numpy",
    "pandas>=3.0.0", # pandas3 required for new time unit handling
    "pypdf",
    "pyYAML",
    "regionmask",
    "ruamel.yaml",
    "scipy",
    "seaborn",
    "smmregrid>=0.1.4",
    # "smmregrid @ git+https://github.com/jhardenberg/smmregrid.git",  # for development
    "typeguard",
    "xarray>=2025.12.0",
    "zarr<=3.1.5",
    # not available on conda
    "pyfdb", #to be pinned when doing operational release
    "gsv-interface==2.13.1"
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx-rtd-theme"
]
notebooks = [
    "ipykernel",
    "healpy"
]
tests = [
    "pytest",
    "pytest-cov",
    "pytest-timeout",
    "pytest-xdist",
    "coverage",
    "cdo"
]
all = [
    "aqua-core[docs]",
    "aqua-core[notebooks]",
    "aqua-core[tests]"
]

[project.scripts]
aqua = "aqua.core.console.main:main"

[project.urls]
homepage = "https://github.com/DestinE-Climate-DT/AQUA/"
documentation = "https://aqua.readthedocs.io/en/latest/index.html"
repository = "https://github.com/DestinE-Climate-DT/AQUA/"
issues = "https://github.com/DestinE-Climate-DT/AQUA/issues/"

[tool.setuptools.packages.find]
where = ["."]
include = ["aqua*"]

# Include non-Python files inside aqua/config and aqua/templates
[tool.setuptools.package-data]
"aqua" = [
    "core/config/**/*",
    "core/templates/**/*"
]

[tool.setuptools.dynamic]
version = {attr = "aqua.core.version.__version__"}

[tool.pytest.ini_options]
# verbose to have full test, run coverage on src folder and use pyproject as a configuration file
addopts = "--durations=10 -v --cov=aqua" 
testpaths = [
    "tests"
]
markers = [
    "aqua: mark test as AQUA framework",
    "slow: tests which are slow",
    "sbatch: tests which require sbatch",
    "graphics: tests which deal with graphics",
    "gsv: mark test that requires ECMWF FDB libraries",
    "catgen: mark test for catalog generator",
    "console: mark test for console",
]

[tool.coverage.run]
branch = true
command_line = '-m pytest tests'
source = ["aqua/core"]
omit = [
    "tests/*", 
    "**/__init__.py",
    ]

